Add files via upload
This bioinformatics project implements a Gibbs Sampling algorithm in Python to discover DNA motifs from the DosR dataset of Mycobacterium tuberculosis (MTB). The focus is on identifying transcription factor binding sites regulated under hypoxic conditions, which are associated with the latent stage of TB infection. This notebook demonstrates motif discovery techniques applied to upstream gene regions and explores the biological significance of the DosR regulon in TB latency and dormancy